Move read-ahead helpers out of the public header
